ブラックリストをInspectionTool側で保持するように修正
#3
