+++ src/main/java/nablarch/intellij/plugin/inspector/JavaOpenApiCatchInspector.kt
@@ -7,7 +7,8 @@ import com.intellij.psi.util.*
 
 class JavaOpenApiCatchInspector(
     private val catchType: PsiType?,
-    private val holder: ProblemsHolder
+    private val holder: ProblemsHolder,
+    private val blacklist: Set<String>
 ) {
 
   fun inspect() {
@@ -15,7 +16,7 @@ class JavaOpenApiCatchInspector(
       is PsiDisjunctionType -> {
         catchType.disjunctions.forEach { type ->
           PsiTypesUtil.getPsiClass(type)?.let {
-            if (!isJavaOpenApi(it)) {
+            if (!isJavaOpenApi(it, blacklist)) {
               addUnpermittedProblem(holder, (type as PsiClassReferenceType).reference)
             }
           }
@@ -23,7 +24,7 @@ class JavaOpenApiCatchInspector(
       }
       else -> {
         PsiTypesUtil.getPsiClass(catchType)?.let {
-          if (!isJavaOpenApi(it)) {
+          if (!isJavaOpenApi(it, blacklist)) {
             addUnpermittedProblem(holder, (catchType as PsiClassReferenceType).reference)
           }
         }
